+++ src/lrcalc/helper/mod.rs
@@ -1,3 +1,2 @@
 pub mod ivector;
 pub mod ivlincomb;
-pub mod part;
