+++ src/lrcalc/mod.rs
@@ -5,7 +5,7 @@ mod lincomb;
 pub mod lrcoef;
 pub mod lriter;
 pub mod optim;
-mod part;
+pub mod part;
 pub mod perm;
 pub mod schublib;
 pub mod schur;
