+++ src/lrcalc/optim.rs
@@ -1,7 +1,5 @@
-use super::helper::{
-    ivector::IntVector,
-    part::{part_entry, part_length, part_leq, part_valid},
-};
+use super::helper::ivector::IntVector;
+use super::part::{part_entry, part_length, part_leq, part_valid};
 
 pub struct SkewShape {
     pub outer: Option<IntVector>,
