+++ src/lrcalc/schur.rs
@@ -4,12 +4,12 @@ use super::helper::{
         ivlc_add_element, ivlc_new, ivlc_new_default, LinearCombination, LC_COPY_KEY, LC_FREE_KEY,
         LC_FREE_ZERO,
     },
-    part::{part_entry, part_valid},
 };
 use super::{
     lrcoef::lrcoef_count,
     lriter::{lrit_expand, lrit_good, lrit_new, lrit_next, LRTableauIterator},
     optim::{optim_coef, optim_fusion, optim_mult, optim_skew},
+    part::{part_entry, part_valid},
 };
 
 pub fn schur_mult(
