+++ src/lrcalc/tests/test_lrskew.rs
@@ -11,8 +11,8 @@ use super::super::{
 pub fn test_schur_lrskew(outer: &IntVector, inner: &IntVector, rows: i32, cols: i32) -> Result<()> {
     let lc = _schur_skew(outer, inner, -1, rows);
 
-    let mut pitr = PartIter::new_box(IntVector::default(rows as u32), rows, cols);
-    while let Some(sh) = pitr.next() {
+    let pitr = PartIter::new_box(IntVector::default(rows as u32), rows, cols);
+    for sh in pitr {
         let coef = _schur_lrcoef(outer, inner, &sh);
         ensure!(coef >= 0, "memory error: sh={:?}", sh.to_vec());
         let expected = if let Some(d) = lc.find(&sh[..]) {
@@ -39,10 +39,10 @@ pub fn test_schur_lrskew(outer: &IntVector, inner: &IntVector, rows: i32, cols: 
 }
 
 pub fn run_test_lrskew(rows: i32, cols: i32) -> Result<()> {
-    let mut pitr1 = PartIter::new_box(IntVector::default(rows as u32), rows, cols);
-    while let Some(p1) = pitr1.next() {
-        let mut pitr2 = PartIter::new_box(IntVector::default(rows as u32), rows, cols);
-        while let Some(p2) = pitr2.next() {
+    let pitr1 = PartIter::new_box(IntVector::default(rows as u32), rows, cols);
+    for p1 in pitr1 {
+        let pitr2 = PartIter::new_box(IntVector::default(rows as u32), rows, cols);
+        for p2 in pitr2 {
             test_schur_lrskew(&p1, &p2, rows, cols)
                 .with_context(|| format!("p1={:?}, p2={:?}", p1.to_vec(), p2.to_vec()))?;
         }
